test: organiza testes de cadastro (models, forms, services)
